+++ RELEASE_NOTES.md
@@ -1,2 +1,3 @@
 v0.80.4
 
+- We now always defer to the HTTP/HTTPS proxy for DNS resolution.